+++ cmd/mq_json/exporter.go
@@ -1,7 +1,7 @@
 package main
 
 /*
-  Copyright (c) IBM Corporation 2016, 2022
+  Copyright (c) IBM Corporation 2016, 2025
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -245,6 +245,9 @@ func Collect() error {
 								if hostname != mqmetric.DUMMY_STRING {
 									pt.Tags["hostname"] = hostname
 								}
+								if showAndSupportsCustomLabel() {
+									pt.Tags["custom"] = mqmetric.GetObjectCustom("", ibmmq.MQOT_Q_MGR)
+								}
 							} else if strings.HasPrefix(key, mqmetric.NativeHAKeyPrefix) {
 								pt.Tags["nha"] = strings.Replace(key, mqmetric.NativeHAKeyPrefix, "", -1)
 								pt.ObjectType = "nha"
@@ -256,7 +259,9 @@ func Collect() error {
 								pt.ObjectType = "queue"
 								pt.Tags["description"] = mqmetric.GetObjectDescription(key, ibmmq.MQOT_Q)
 								pt.Tags["cluster"] = mqmetric.GetQueueAttribute(key, ibmmq.MQCA_CLUSTER_NAME)
-
+								if showAndSupportsCustomLabel() {
+									pt.Tags["custom"] = mqmetric.GetObjectCustom(key, ibmmq.MQOT_Q)
+								}
 							}
 							addMetaLabels(pt.Tags)
 						}
@@ -345,6 +350,9 @@ func Collect() error {
 									pt.Tags["queue"] = qName
 									pt.Tags["usage"] = usageString
 									pt.Tags["description"] = mqmetric.GetObjectDescription(qName, ibmmq.MQOT_Q)
+									if showAndSupportsCustomLabel() {
+										pt.Tags["custom"] = mqmetric.GetObjectCustom(qName, ibmmq.MQOT_Q)
+									}
 									pt.Tags["cluster"] = mqmetric.GetQueueAttribute(key, ibmmq.MQCA_CLUSTER_NAME)
 									pt.Tags["platform"] = platformString
 									addMetaLabels(pt.Tags)
@@ -400,6 +408,9 @@ func Collect() error {
 									pt.Tags["qmgr"] = strings.TrimSpace(qMgrName)
 									pt.Tags["platform"] = platformString
 									pt.Tags["description"] = mqmetric.GetObjectDescription("", ibmmq.MQOT_Q_MGR)
+									if showAndSupportsCustomLabel() {
+										pt.Tags["custom"] = mqmetric.GetObjectCustom("", ibmmq.MQOT_Q_MGR)
+									}
 									hostname := mqmetric.GetQueueManagerAttribute(config.cf.QMgrName, ibmmq.MQCACF_HOST_NAME)
 									if hostname != mqmetric.DUMMY_STRING {
 										pt.Tags["hostname"] = hostname
@@ -694,4 +705,8 @@ func addMetaLabels(tags map[string]string) {
 			tags[config.cf.MetadataTagsArray[i]] = config.cf.MetadataValuesArray[i]
 		}
 	}
+}
+
+func showAndSupportsCustomLabel() bool {
+	return config.cf.CC.ShowCustomAttribute
 }