+++ cmd/mq_opentsdb/exporter.go
@@ -1,7 +1,7 @@
 package main
 
 /*
-  Copyright (c) IBM Corporation 2016,2022
+  Copyright (c) IBM Corporation 2016,2025
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -247,6 +247,9 @@ func Collect() error {
 								if hostname != mqmetric.DUMMY_STRING {
 									tags["hostname"] = hostname
 								}
+								if showAndSupportsCustomLabel() {
+									tags["custom"] = mqmetric.GetObjectCustom("", ibmmq.MQOT_Q_MGR)
+								}
 							} else if strings.HasPrefix(key, mqmetric.NativeHAKeyPrefix) {
 								series = "nha"
 								tags["nha"] = strings.Replace(key, mqmetric.NativeHAKeyPrefix, "", -1)
@@ -265,7 +268,9 @@ func Collect() error {
 								tags["usage"] = usage
 								tags["description"] = mqmetric.GetObjectDescription(key, ibmmq.MQOT_Q)
 								tags["cluster"] = mqmetric.GetQueueAttribute(key, ibmmq.MQCA_CLUSTER_NAME)
-
+								if showAndSupportsCustomLabel() {
+									tags["custom"] = mqmetric.GetObjectCustom(key, ibmmq.MQOT_Q)
+								}
 							}
 							addMetaLabels(tags)
 
@@ -351,6 +356,9 @@ func Collect() error {
 						tags["usage"] = usage
 						tags["description"] = mqmetric.GetObjectDescription(key, ibmmq.MQOT_Q)
 						tags["cluster"] = mqmetric.GetQueueAttribute(key, ibmmq.MQCA_CLUSTER_NAME)
+						if showAndSupportsCustomLabel() {
+							tags["custom"] = mqmetric.GetObjectCustom(key, ibmmq.MQOT_Q)
+						}
 						addMetaLabels(tags)
 
 						f := mqmetric.QueueNormalise(attr, value.ValueInt64)
@@ -472,6 +480,9 @@ func Collect() error {
 						hostname := mqmetric.GetQueueManagerAttribute(config.cf.QMgrName, ibmmq.MQCACF_HOST_NAME)
 						if hostname != mqmetric.DUMMY_STRING {
 							tags["hostname"] = hostname
+						}
+						if showAndSupportsCustomLabel() {
+							tags["custom"] = mqmetric.GetObjectCustom("", ibmmq.MQOT_Q_MGR)
 						}
 						addMetaLabels(tags)
 
@@ -706,4 +717,8 @@ func addMetaLabels(tags map[string]string) {
 			tags[config.cf.MetadataTagsArray[i]] = config.cf.MetadataValuesArray[i]
 		}
 	}
+}
+
+func showAndSupportsCustomLabel() bool {
+	return config.cf.CC.ShowCustomAttribute
 }