+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 Newest updates are at the top of this file.
 
+### xxx xx 2025 (v5.6.5)
+* Update to MQ 9.4.4
+* Add showCustomAttribute filter config setting to include the CUSTOM attribute as metric tag/label
+
 ### Jun 19 2025 (no new version)
 * Improve container building
 
@@ -19,6 +23,7 @@ Newest updates are at the top of this file.
 ### Unreleased
 * Fix boolean logic for `isFirstCollection` in `mq_prometheus (#385)`
   * Ensure proper collection on the first poll and at regular intervals thereafter
+* Add showCustomAttribute filter config setting to include the CUSTOM attribute as metric tag/label
 
 ### Feb 28 2025 (v5.6.2)
 * Update to MQ 9.4.2